+++ compiler/src/CLI_errors.ml
@@ -57,4 +57,9 @@ let check_options () =
   if !latexfile <> ""
   then warning Deprecated Location.i_dummy
          "the [-latex] option has been deprecated since March 2023; use [jazz2tex] instead";
+
+  if Option.is_some !ct_list || !infer
+  then warning Deprecated Location.i_dummy
+         "the command-line options for constant-time have been deprecated since April 2024; use [jazzcheck] instead";
+
   List.iter chk_out_file [ outfile; latexfile; ecfile ]