+++ CHANGELOG.md
@@ -78,6 +78,11 @@
 - Extraction as EasyCrypt code targets version 2024.01
   ([PR #690](https://github.com/jasmin-lang/jasmin/pull/690)).
 
+- The Constant-Time security checker is now available as a separate `jazzcheck`
+  tool; the `-checkCT`, `-checkCTon`, and `-infer` command line options are
+  deprecated
+  ([PR #766](https://github.com/jasmin-lang/jasmin/pull/766)).
+
 ## Bug fixes
 
 - Type-checking rejects invalid variants of primitive operators